--- src/lib/Libifl/int_deleteput.c
@@ -1,110 +1,0 @@
-/*
- * Copyright (C) 1994-2020 Altair Engineering, Inc.
- * For more information, contact Altair at www.altair.com.
- *
- * This file is part of both the OpenPBS software ("OpenPBS")
- * and the PBS Professional ("PBS Pro") software.
- *
- * Open Source License Information:
- *
- * OpenPBS is free software. You can redistribute it and/or modify it under
- * the terms of the GNU Affero General Public License as published by the
- * Free Software Foundation, either version 3 of the License, or (at your
- * option) any later version.
- *
- * OpenPBS is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public
- * License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * Commercial License Information:
- *
- * PBS Pro is commercially licensed software that shares a common core with
- * the OpenPBS software.  For a copy of the commercial license terms and
- * conditions, go to: (http://www.pbspro.com/agreement.html) or contact the
- * Altair Legal Department.
- *
- * Altair's dual-license business model allows companies, individuals, and
- * organizations to create proprietary derivative works of OpenPBS and
- * distribute them - whether embedded or bundled with other software -
- * under a commercial license agreement.
- *
- * Use of Altair's trademarks, including but not limited to "PBS™",
- * "OpenPBS®", "PBS Professional®", and "PBS Pro™" and Altair's logos is
- * subject to Altair's trademark licensing policies.
- */
-
-/**
- * @file	int_manage2.c
- * @brief
- * 	The send-request side of the PBS_manager function
- */
-
-#include <pbs_config.h>   /* the master config generated by configure */
-
-#include <string.h>
-#include <stdio.h>
-#include "libpbs.h"
-#include "dis.h"
-#include "net_connect.h"
-#include "tpp.h"
-
-
-/**
- * @brief
- *      -encode a Delete Batch Request
- *
- * @par Functionality:
- *              This request is used for delete operation.
- *              
- *
- * @param[in] c - socket descriptor
- * @param[in] command - command type
- * @param[in] objtype - object type
- * @param[in] objname - object name
- * @param[in] aoplp - pointer to attropl structure(list)
- * @param[in] prot - PROT_TCP or PROT_TPP
- * @param[in] msgid - message id
- *
- * @return      int
- * @retval      DIS_SUCCESS(0)  success
- * @retval      error code      error
- *
- */
-int
-PBSD_delete_put(int c, int function, int command, int objtype, char **objnames, struct attropl *aoplp, char *extend, int prot, char **msgid)
-{
-	int rc;
-
-	if (prot == PROT_TCP) {
-		DIS_tcp_funcs();
-	} else {
-		if ((rc = is_compose_cmd(c, IS_CMD, msgid)) != DIS_SUCCESS)
-			return rc;
-	}
-
-	if ((rc = encode_DIS_ReqHdr(c, function, pbs_current_user)) ||
-		(rc = encode_DIS_Delete(c, command, objtype, objnames, aoplp)) ||
-		(rc = encode_DIS_ReqExtend(c, extend))) {
-		if (prot == PROT_TCP) {
-			if (set_conn_errtxt(c, dis_emsg[rc]) != 0)
-				return (pbs_errno = PBSE_SYSTEM);
-		}
-		return (pbs_errno = PBSE_PROTOCOL);
-	}
-
-	if (prot == PROT_TPP) {
-		pbs_errno = PBSE_NONE;
-		if (dis_flush(c))
-			pbs_errno = PBSE_PROTOCOL;
-		return pbs_errno;
-	}
-
-	if (dis_flush(c)) {
-		return (pbs_errno = PBSE_PROTOCOL);
-	}
-	return 0;
-}
